+++ client/components/forms/drag-drop/index.tsx
@@ -1,3 +1,4 @@
+import './index.scss';
 import React, { useCallback } from 'react';
 import { useDropzone } from 'react-dropzone';