+++ client/components/forms/scrollbar.tsx
@@ -1,0 +1,2 @@
+import '@cstyles/scrollbar';
+export { Scrollbar } from 'react-scrollbars-custom';