+++ client/components/index.tsx
@@ -10,4 +10,5 @@ export * from './forms/label-selector';
 export * from './forms/number-input';
 export * from './forms/messages';
 export * from './forms/drag-drop';
+export * from './forms/scrollbar';
 export * from './misc/byte-unit';