+++ client/stores/selectors/columns.ts
@@ -1,22 +1,9 @@
 import { RootState } from '../state';
 import { createSelector } from '@reduxjs/toolkit';
-import { TorrentFields } from '@client/models';
-
-export const selectAllColumns =
-  createSelector(
-    [(state: RootState) => state.settings.columns.entries],
-    entries => Object.entries(entries)
-      .map(([type, entry]) => ({
-        type: Number(type) as TorrentFields,
-        ...entry
-      }))
-  )
+import { ColumnType } from '@client/stores';
 
 export const selectColumnSettings =
   (state: RootState) => state.settings.columns
-
-export const selectColumnOrder =
-  (state: RootState) => state.settings.columns.order
 
 export const selectColumns = createSelector(
   [(state: RootState) => state.settings.columns,