+++ client/stores/torrents-store/actions.ts
@@ -1,6 +1,8 @@
 import { createAction } from '@reduxjs/toolkit';
 
-import { Torrent, TorrentFields } from '@client/models';
+import { ColumnType } from '../settings-store/state';
+
+import { Torrent } from '@client/models';
 import { TransmissionPriorityType as TorrentPriority } from '@transmission';
 
 /** action for when we're updating one or more torrents. */
@@ -12,7 +14,7 @@ export const torrentsAdded = createAction<{ torrents: Torrent[] }>('torrents/add
 /** action for when one or more torrents have been removed. */
 export const torrentsRemoved = createAction<{ ids: number[] }>('torrents/removed')
 
-export const activeFieldChanged = createAction<{ field: TorrentFields }>('settings/column-select')
+export const activeFieldChanged = createAction<{ field: ColumnType }>('settings/column-select')
 
 // updaters for specific fields in a torrent.
 export const torrentPriorityChanged = createAction<{ ids: number[], priority: TorrentPriority }>('torrents/priority-updated')